fix(review): extract shared scoring to search-utils, sort before limit
Address PR review comments:
1. Extract duplicated scoring logic from client.ts and mock-api.ts into
   shared search-utils.ts utility
2. Sort merged results before applying the limit cap so the best matches
   are never excluded by truncation
