+++ web-app/src/api/client.ts
@@ -9,6 +9,7 @@ import {
   validateResponse,
 } from './validation'
 import { mockApi } from './mock-api'
+import { scoreNameMatch } from './search-utils'
 
 import { calendarApi } from '@/features/assignments/api/calendar-client'
 
@@ -202,54 +203,6 @@ async function apiRequest<T>(
       `${method} ${endpoint}: Invalid JSON response (Content-Type: ${contentType || 'unknown'}, status: ${response.status})`
     )
   }
-}
-
-/**
- * Normalize a string for accent-insensitive comparison.
- * Uses Unicode NFD decomposition to strip combining diacritical marks,
- * e.g. "Bühler" → "buhler", "Renée" → "renee".
- */
-function normalizeForComparison(s: string): string {
-  return s
-    .normalize('NFD')
-    .replace(/[\u0300-\u036f]/g, '')
-    .toLowerCase()
-}
-
-/** Match scores for ranking person search results (higher = better match). */
-const SCORE_EXACT = 4
-const SCORE_PREFIX = 3
-const SCORE_QUERY_CONTAINS_FIELD = 2
-const SCORE_SUBSTRING = 1
-
-/**
- * Score how well a single field matches a query term.
- * Higher score = better match.
- */
-function scoreField(field: string, query: string): number {
-  if (field === query) return SCORE_EXACT
-  if (field.startsWith(query)) return SCORE_PREFIX
-  if (query.startsWith(field)) return SCORE_QUERY_CONTAINS_FIELD
-  if (field.includes(query)) return SCORE_SUBSTRING
-  return 0
-}
-
-/**
- * Score how well a person's name matches two search terms, trying both orderings.
- * Used to re-rank merged results so "Lastname Firstname" and "Firstname Lastname"
- * both surface the correct person first.
- */
-function scoreNameMatch(person: PersonSearchResult, term1: string, term2: string): number {
-  const first = normalizeForComparison(person.firstName ?? '')
-  const last = normalizeForComparison(person.lastName ?? '')
-  const t1 = normalizeForComparison(term1)
-  const t2 = normalizeForComparison(term2)
-
-  // Try both orderings: t1=firstName/t2=lastName and t1=lastName/t2=firstName
-  const score1 = scoreField(first, t1) + scoreField(last, t2)
-  const score2 = scoreField(first, t2) + scoreField(last, t1)
-
-  return Math.max(score1, score2)
 }
 
 /** Properties requested from the Elasticsearch person search endpoint */
@@ -459,9 +412,9 @@ export const api = {
         makeRequest(lastName, firstName),
       ])
 
-      // Merge, deduplicate, and re-rank results so the best name match appears first.
-      // Without re-ranking, the "original" ordering's results always come first,
-      // which produces poor results when the user types last name before first name.
+      // Merge and deduplicate results from both orderings, then re-rank so the
+      // best name match appears first. Sorting happens before the limit cap so
+      // that the best matches are never excluded by an arbitrary truncation point.
       const limit = options?.limit ?? DEFAULT_SEARCH_RESULTS_LIMIT
       const seen = new Set<string>()
       const merged: PersonSearchResult[] = []
@@ -470,17 +423,19 @@ export const api = {
         if (id && !seen.has(id)) {
           seen.add(id)
           merged.push(item)
-          if (merged.length >= limit) break
         }
       }
 
       // Re-rank: score each result by how well its firstName/lastName match the
       // two search terms (trying both orderings), then sort best matches first.
       merged.sort((a, b) => {
-        return scoreNameMatch(b, firstName, lastName) - scoreNameMatch(a, firstName, lastName)
+        const scoreA = scoreNameMatch(a.firstName ?? '', a.lastName ?? '', firstName, lastName)
+        const scoreB = scoreNameMatch(b.firstName ?? '', b.lastName ?? '', firstName, lastName)
+        return scoreB - scoreA
       })
 
-      return { items: merged, totalItemsCount: merged.length }
+      const capped = merged.slice(0, limit)
+      return { items: capped, totalItemsCount: capped.length }
     }
 
     // Single-term search: search both firstName and lastName fields
